+++ scripts/prepare_docs.bash
@@ -166,6 +166,12 @@ echo -e "~~~text\n$(odoo-helper odoo --help)\n~~~\n\n" >> $PROJECT_DIR/docs/comm
 echo -e "#### odoo-helper odoo recompute\n\n" >> $PROJECT_DIR/docs/command-reference.md;
 echo -e "~~~text\n$(odoo-helper odoo recompute --help)\n~~~\n\n" >> $PROJECT_DIR/docs/command-reference.md;
 
+echo -e "#### odoo-helper odoo recompute-menu\n\n" >> $PROJECT_DIR/docs/command-reference.md;
+echo -e "~~~text\n$(odoo-helper odoo recompute-menu --help)\n~~~\n\n" >> $PROJECT_DIR/docs/command-reference.md;
+
+echo -e "#### odoo-helper odoo clean-compiled-assets\n\n" >> $PROJECT_DIR/docs/command-reference.md;
+echo -e "~~~text\n$(odoo-helper odoo clean-compiled-assets --help)\n~~~\n\n" >> $PROJECT_DIR/docs/command-reference.md;
+
 echo -e "### odoo-helper tr\n\n" >> $PROJECT_DIR/docs/command-reference.md;
 echo -e "~~~text\n$(odoo-helper tr --help)\n~~~\n\n" >> $PROJECT_DIR/docs/command-reference.md;